Added a confirm to my delete event listener in case I ever click delete by accident

+++ src/scripts/journal.js
@@ -23,7 +23,6 @@ document.querySelector("#btn-container").addEventListener("click", () => {
 
 //Keyup to set character limits on the Concepts Covered and Journal Entry portion of the form
 document.querySelector("#output-container").addEventListener("keyup", () => {
-    console.log(document.querySelector("#journalEntry").value.length)
     if (document.querySelector("#conceptsCovered").value.length >= 100) {
         alert("You have reached the character limit for this input")
     } else if (document.querySelector("#journalEntry").value.length >= 200) {
@@ -38,8 +37,13 @@ document.querySelector("#output-container").addEventListener("click", () => {
     //POSTS journal entry to journal.json if record button is clicked
     if (event.target.id.includes("record")) {
         // Creates a new journal entry based on the information in the form and posts it to journal.json then reprints the entries to the DOM
-        const curseArray = ["fuck", "shit", "ass", "cocksucker", "damn"]
+
+        // Takes the each word in concepts covered and journal entry, stores it in inputArray, and tests each word against the words in curseArray.
+        const curseArray = ["fuck", "shit", "ass", "cocksucker", "damn", "motherfucker", "fucker"]
         const inputArray = document.querySelector("#journalEntry").value.split(" ")
+        document.querySelector("#conceptsCovered").value.split(" ").forEach(word => {
+            inputArray.push(word)
+        });
 
         //Checks if any word in my inputArray matches any word in my curseArray and returns an alert, else it posts the entry and prints the journal to the DOM
         if (inputArray.some(word => curseArray.includes(word))) {
@@ -50,11 +54,13 @@ document.querySelector("#output-container").addEventListener("click", () => {
         //If the delete button is clicked the entry is deleted
     } else if (event.target.id.includes("delete-btn")) {
 
-        // Grabs the ID number from the ID and stores it in IDToDelete
-        const IDToDelete = event.target.id.split("-")[2]
+        if (confirm("You are about to delete a jounral entry")) {
+            // Grabs the ID number from the ID and stores it in IDToDelete
+            const IDToDelete = event.target.id.split("-")[2]
 
-        // Deletes the JSON object with the chosen ID then reprints the entries to the DOM
-        APIManager.deleteJournalEntry(IDToDelete).then(DOMPrinter.printJournalToTheDOM)
+            // Deletes the JSON object with the chosen ID then reprints the entries to the DOM
+            APIManager.deleteJournalEntry(IDToDelete).then(DOMPrinter.printJournalToTheDOM)
+        } 
 
         //If the edit button is clicked the form is injected into that journal entry with prepopulated information from the journal entry
     } else if (event.target.id.includes("edit-btn")) {
